# Improve performance of Python dependency inference