--- a/src/python/pants/backend/python/dependency_inference/BUILD
+++ b/src/python/pants/backend/python/dependency_inference/BUILD
@@ -27,7 +27,6 @@ python_tests(
   name="tests",
   dependencies=[
     ":dependency_inference",
-    "src/python/pants/backend/project_info",
     "src/python/pants/testutil:test_base",
   ],
   tags = {"type_checked"},