--- a/src/python/pants/backend/python/dependency_inference/module_mapper.py
+++ b/src/python/pants/backend/python/dependency_inference/module_mapper.py
@@ -3,10 +3,10 @@
 
 from dataclasses import dataclass
 from pathlib import PurePath
-from typing import Dict, Iterator, Optional
+from typing import Dict, Optional
 
 from pants.backend.python.target_types import PythonRequirementsField, PythonSources
-from pants.base.specs import AddressSpecs, AscendantAddresses, DescendantAddresses
+from pants.base.specs import AddressSpecs, DescendantAddresses
 from pants.core.util_rules.strip_source_roots import (
     SourceRootStrippedSources,
     StripSourcesFieldRequest,
@@ -16,7 +16,6 @@ from pants.engine.collection import DeduplicatedCollection
 from pants.engine.rules import rule
 from pants.engine.selectors import Get, MultiGet
 from pants.engine.target import Targets
-from pants.source.source_root import AllSourceRoots
 from pants.util.frozendict import FrozenDict
 
 
@@ -31,42 +30,44 @@ class PythonModule:
         )
         return cls(module_name_with_slashes.as_posix().replace("/", "."))
 
-    @property
-    def name_as_path(self) -> PurePath:
-        return PurePath(self.module.replace(".", "/"))
 
-    def address_spec(self, *, source_root: str) -> AscendantAddresses:
-        """The spec for all candidate targets which could feasibly own the module.
+@dataclass(frozen=True)
+class FirstPartyModuleToAddressMapping:
+    mapping: FrozenDict[str, Address]
 
-        This uses AscendantAddresses because targets can own files in subdirs (e.g. rglobs). We also
-        use the package path, e.g. `helloworld/util/__init__.py`, rather than the module path to
-        ensure that we capture all possible targets. It is okay if this directory does not actually
-        exist.
-        """
-        return AscendantAddresses(directory=str(PurePath(source_root) / self.name_as_path))
+    def address_for_module(self, module: str) -> Optional[Address]:
+        target = self.mapping.get(module)
+        if target is not None:
+            return target
+        # If the module is not found, try the parent, if any. This is to accommodate `from`
+        # imports, where we don't care about the specific symbol, but only the module. For example,
+        # with `from typing import List`, we only care about `typing`.
+        # Unlike with third party modules, we do not look past the direct parent.
+        if "." not in module:
+            return None
+        parent_module = module.rsplit(".", maxsplit=1)[0]
+        return self.mapping.get(parent_module)
 
-    def possible_stripped_paths(self) -> Iterator[PurePath]:
-        """Given a module like `helloworld.util`, convert it back to its possible paths.
 
-        Each module has either 2 or 4 possible paths. For example, given the module
-        `helloworld.util`:
-
-        - helloworld/util.py
-        - helloworld/util/__init__.py
-        - helloworld.py
-        - helloworld/__init__.py
-
-        The last two possible paths look at the parent module, if any. This is to accommodate `from`
-        imports, where we don't care about the specific symbol, but only the module. For example,
-        with `from typing import List`, we only care about `typing`.
-        """
-        module_name_with_slashes = PurePath(self.module.replace(".", "/"))
-        yield self.name_as_path.with_suffix(".py")
-        yield self.name_as_path / "__init__.py"
-        parent = module_name_with_slashes.parent
-        if str(parent) != ".":
-            yield parent.with_suffix(".py")
-            yield parent / "__init__.py"
+@rule
+async def map_first_party_modules_to_addresses() -> FirstPartyModuleToAddressMapping:
+    all_targets = await Get[Targets](AddressSpecs([DescendantAddresses("")]))
+    candidate_targets = tuple(tgt for tgt in all_targets if tgt.has_field(PythonSources))
+    sources_per_target = await MultiGet(
+        Get[SourceRootStrippedSources](StripSourcesFieldRequest(tgt[PythonSources]))
+        for tgt in candidate_targets
+    )
+    modules_to_addresses: Dict[str, Address] = {}
+    for tgt, sources in zip(candidate_targets, sources_per_target):
+        for f in sources.snapshot.files:
+            module = PythonModule.create_from_stripped_path(PurePath(f)).module
+            # NB: If >1 targets have the same module, we do not record the module. This is to
+            # avoid ambiguity.
+            if module in modules_to_addresses:
+                modules_to_addresses.pop(module)
+            else:
+                modules_to_addresses[module] = tgt.address
+    return FirstPartyModuleToAddressMapping(FrozenDict(modules_to_addresses))
 
 
 @dataclass(frozen=True)
@@ -110,31 +111,23 @@ class PythonModuleOwners(DeduplicatedCollection[Address]):
 
 
 @rule
-async def map_python_module_to_targets(
+async def map_module_to_addresses(
     module: PythonModule,
-    source_roots: AllSourceRoots,
+    first_party_mapping: FirstPartyModuleToAddressMapping,
     third_party_mapping: ThirdPartyModuleToAddressMapping,
 ) -> PythonModuleOwners:
     third_party_address = third_party_mapping.address_for_module(module.module)
     if third_party_address:
         return PythonModuleOwners([third_party_address])
-    unfiltered_candidate_targets = await Get[Targets](
-        AddressSpecs(module.address_spec(source_root=src_root.path) for src_root in source_roots)
-    )
-    candidate_targets = tuple(
-        tgt for tgt in unfiltered_candidate_targets if tgt.has_field(PythonSources)
-    )
-    sources_per_target = await MultiGet(
-        Get[SourceRootStrippedSources](StripSourcesFieldRequest(tgt[PythonSources]))
-        for tgt in candidate_targets
-    )
-    candidate_files = {str(p) for p in module.possible_stripped_paths()}
-    return PythonModuleOwners(
-        tgt.address
-        for tgt, sources in zip(candidate_targets, sources_per_target)
-        if bool(candidate_files.intersection(sources.snapshot.files))
-    )
+    first_party_address = first_party_mapping.address_for_module(module.module)
+    if first_party_address:
+        return PythonModuleOwners([first_party_address])
+    return PythonModuleOwners()
 
 
 def rules():
-    return [map_python_module_to_targets, map_third_party_modules_to_addresses]
+    return [
+        map_first_party_modules_to_addresses,
+        map_third_party_modules_to_addresses,
+        map_module_to_addresses,
+    ]